Changes in response to code review and other improvements.
